reduce size of the chat file by splitting the code multiples files to help readibility
